Created custom hook for addprod icon drop down. file in hooks folder, imported into component in AddIconDropDown and exported into Invetory. working on breaking down the remaining code in inventory
