+++ client/src/components/Inventory/Inventory.jsx
@@ -9,12 +9,13 @@ import OrderNowPopup from './popups/OrderNow';
 import IncomingPopup from './popups/IncomingOrders';
 import './inventory.css';
 import './popups/popup.css';
-import './dropdown.css';
+import './AddIconDropDown.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faFile, faSquarePlus, faCloudArrowUp, } from '@fortawesome/free-solid-svg-icons';
+import { faFile } from '@fortawesome/free-solid-svg-icons';
 import { sendCSVfile } from '../../services/inventoryAPIcalls';
 import { authenticateUser } from '../../services/authenticationAPIcalls';
 import { useQuery } from 'react-query';
+import DropDownIcon from './AddIconDropDown.jsx'
 
 export default function Inventory({ tempInStock }) {
   /*   const { data } = useQuery('authenticateUser', authenticateUser);
@@ -24,36 +25,8 @@ export default function Inventory({ tempInStock }) {
   const [itemId, setItemId] = useState(0);
   // this is the whole product object to be passed down into popup
   const [productForPopup, setProductForPopup] = useState('');
-  const [isDropOpen, setIsDropOpen] = useState(false);
-  const dropdownRef = useRef(null);
   const [dragInventory, setDragInventory] = useState([]);
 
-  //-------------- Icon Drop down for add product -----------------
-  const toggleDropdown = () => {
-    setIsDropOpen(!isDropOpen);
-  };
-
-  // closes drop down after user selects an item from the menu
-  const handleDropClose = () => {
-    setIsDropOpen(false);
-  };
-
-  // close dropdown if user clicks outside of the menu
-
-  const handleClickOutside = (event) => {
-    if (dropdownRef.current && !dropdownRef.current.contains(event.target)) {
-      setIsDropOpen(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener('click', handleClickOutside, true);
-    return () => {
-      document.removeEventListener('click', handleClickOutside, true);
-    };
-  }, []);
-
-  // -------------------- load and reload inventory ------------------------------
 
   useEffect(() => {
     console.log(inventory);
@@ -65,15 +38,6 @@ export default function Inventory({ tempInStock }) {
 
 
   // ------------- update items' input values when user changes them ---------------
-
-  // -------------------------- CSV ----------------------------
-  const handleChange = async (e) => {
-    console.log(e.target.files[0]);
-    await sendCSVfile(e.target.files[0]);
-    handleDropClose();
-    reloadInventory(); // not working yet
-  };
-
   // ------------- update items' input values  ---------------
 
   const handleKeyDown = async (event, id, field, value) => {
@@ -146,18 +110,15 @@ export default function Inventory({ tempInStock }) {
 
   // -------------------------- drag and drop --------------------
   const handleDragEnd = (result) => {
-    return result.destination
-      ? (() => {
-        const startIndex = result.source.index;
-        const endIndex = result.destination.index;
+    if (!result.destination) {
+      return;
+    }
 
-        const newInventory = Array.from(dragInventory);
-        const [removed] = newInventory.splice(startIndex, 1);
-        newInventory.splice(endIndex, 0, removed);
+    const items = Array.from(dragInventory);
+    const [reorderedItem] = items.splice(result.source.index, 1);
+    items.splice(result.destination.index, 0, reorderedItem);
 
-        setDragInventory(newInventory);
-      })()
-      : null;
+    setDragInventory(items);
   };
 
   // ----------------------------------------------------------
@@ -179,55 +140,7 @@ export default function Inventory({ tempInStock }) {
               </td>
             ))}
             <td id="add-prod-td">
-              <div className="dropdown-icon">
-                <button className="addprodicon">
-                  <FontAwesomeIcon
-                    icon={faSquarePlus}
-                    onClick={toggleDropdown}
-                  />
-                </button>
-                {isDropOpen && (
-                  <div ref={dropdownRef} className="dropdown-menu">
-                    <ul>
-                      <li>
-                        <a
-                          onClick={() => {
-                            handleDisplayRow();
-                            handleDropClose();
-                            document
-                              .getElementById('scrollForAddRow')
-                              .scrollIntoView({ behavior: 'smooth' });
-                          }}
-                        >
-                          <FontAwesomeIcon
-                            icon={faSquarePlus}
-                            className="fa-dropdown"
-                          />
-                          Add Product
-                        </a>
-                      </li>
-                      <li>
-                        <a>
-                          <label>
-                            <FontAwesomeIcon
-                              icon={faCloudArrowUp}
-                              className="fa-dropdown"
-                            />
-                            From file
-                            <input
-                              type="file"
-                              accept=".csv"
-                              onChange={(e) => handleChange(e)}
-                              // onClick={handleDropClose}
-                              style={{ display: 'none' }}
-                            />
-                          </label>
-                        </a>
-                      </li>
-                    </ul>
-                  </div>
-                )}
-              </div>
+                <DropDownIcon />
             </td>
           </tr>
         </thead>
@@ -261,6 +174,7 @@ export default function Inventory({ tempInStock }) {
                           ref={provided.innerRef}
                           {...provided.draggableProps}
                           {...provided.dragHandleProps}
+                          className={snapshot.isDragging ? "dragging" : ""}
                         >
                           <td id="scrollForAddRow" className="item-sku">
                             {/* this id catches the scrollintoview when clicking add product */}
