+++ client/src/services/inventoryAPIcalls.js
@@ -35,7 +35,7 @@ export async function createInventoryItem(product) {
   const response = await fetch(`${API_URL}/inventory/`, {
     method: 'POST',
     body: JSON.stringify({
-      sku: product.sku.toUppercase(),
+      sku: product.sku,
       brand: product.brand,
       productName: product.productName,
       description: product.description,
